update gatsby 2.32.6 -> 4.10.3
After reviewing Dependabot alerts[0], an update to newer Gatsby was
called for, and this requires a number of Gatsby related packages as
well.

I was unable to find a set of dependencies which would work together
with Gatsby 3. However, the latest release[1] of Gatsby is 4.10 and I
was able to get packages to work for this version.

Testing with this revealed a display issue. The font size of the main
navigation bar at the top of the pages is not being set correctly.

This introduces a horizontal scrollbar, and the language switching
between NL and EN is scrolled off to the right.

Other than the display issue, the rest of the site seems to render well
enough, even though fonts in various places have different sizes.

One of the packages which is upgraded is @commonground/design-system;
this may be at the heart of the differences in display.  This upgrade
seems like quite a jump (from 12.5 to 21.1) and while an intermediate
version may work, a better approach would be to figure out how to get
the display elements to be correct with the latest version.

[0] https://github.com/Signalen/signalen.org/security/dependabot
[1] https://github.com/gatsbyjs/gatsby/releases